feat: add general ledger MVP and harden accounting, exports, sessions, and UI (#126)
* feat: add general ledger MVP and harden accounting, exports, sessions, and UI

General ledger (Buku Besar) built on the existing journal model:
- getGeneralLedger with per-account running balance (opening balance carried
  into the visible range), GET /api/reports/general-ledger, and a new page
  under Laporan with date/account filters; bounded at 5,000 rows + truncated
  flag. Data model needed no schema change.
- Frontend wiring: api client, query keys + invalidation, route, nav item.

Accounting/API hardening (bug hunt):
- Reject fractional rupiah amounts instead of silently rounding (service +
  zod schema now require an integer) and enforce the IDR ceiling.
- Idempotency race: a UNIQUE constraint hit by two parallel POSTs with the
  same key now resolves to a replay of the winner instead of a 500.
- Session token rotation keeps the rotated-out hash valid for a 60s grace
  window (migration 0005) so parallel in-flight requests with the old cookie
  are not logged out at the 7-day rotation boundary.
- CSV export without a status filter now exports all statuses (posted +
  voided), matching the "Semua status" list view.

UI consistency pass:
- Unify page section rhythm (space-y-4), fix ineffective col-span grid
  placement on account/transaction toolbars, remove double card padding on
  auth/404 pages, align list row paddings.
- Normalize heading hierarchy: card titles h2, single h1 per page, brand
  panel copy no longer introduces an h2 before the page h1.
- scripts/ui-audit.mjs gains design-consistency checks (heading outline +
  equal column widths on uniform grids).

Cleanup & docs:
- Remove dead pre-MVP files (unused UI components, service worker, benchmark
  seed script, audit JSON artifacts, dead CSS) and stale SEO copy; fix
  CODEOWNERS.
- Re-sync documentation (accounting rules, permission matrix, tenant
  isolation, testing, OpenAPI, READMEs) and CHANGELOG to MVP scope.

Verified: 250 unit/integration tests, typecheck, lint, build, full Playwright
UI audit (26 page x viewport combos), and live HTTP probes for the idempotency
race, rotation grace, and CSV status default.


* fix: send Origin header on UI audit API POSTs to satisfy CSRF

The standalone UI audit script authenticates and seeds transactions via
context.request, which omits the Origin header. Against hardened
deployments (staging/prod) the worker's CSRF origin check (ADR 0003)
rejects cookie-authenticated POSTs without an allowed Origin with 403
csrf_invalid, silently skipping the transaction-detail route. Send a
browser-like Origin header so the audit covers every page in any
environment, not just local builds with relaxed CSRF.


* feat: add Trakteer support link across app chrome

Ledjer has no payment gateway, so add a voluntary support link to
Trakteer. A single source-of-truth URL (src/lib/external.ts) feeds a
footer link on public pages (landing, login, register) and a heart-icon
row in the authenticated sidebar and mobile drawer, reaching every page
without per-page markup. Opens in a new tab with rel=noopener.


* docs: sync READMEs with current project state

Update docs that drifted from the codebase: list the general-ledger
report and Google OAuth in the root feature set plus the OAuth secrets,
correct the dev-server note (the Worker runs same-origin via the
@cloudflare vite plugin, not a separate 8788 proxy), fix the zod version
(v4), describe authenticated e2e coverage, and advance the D1 migration
range to 0001-0005.


* fix: resolve SonarCloud findings in UI audit script and GL report

Fix the 13 new-code issues flagged on PR analysis: drop the dead
console/page/failed-request listeners in auditViewport (S4030), remove
the extra argument passed to auditRoute (S930), replace the hardcoded
/tmp output dir with os.tmpdir() (S5443), drop Math.random in favor of
node:crypto randomUUID (S2245), flatten the design-check functions to
cut cognitive complexity (S3776), and apply the minor style rules
(Number.parseFloat/parseInt, spread-free for..of, Array.at, optional
chaining).


---------

+++ apps/web/worker/routes/transactions.routes.ts
@@ -22,7 +22,7 @@ const postTransactionSchema = z.object({
   transactionDate: dateSchema,
   cashAccountId: z.string().min(1),
   counterAccountId: z.string().min(1),
-  amountIdr: z.number().positive(),
+  amountIdr: z.number().int().positive(),
   description: z.string().min(1).max(200),
   idempotencyKey: z.string().min(8).max(160),
 });
